Update sample with some workarounds.
